+++ R/imports.R
@@ -1,2 +1,8 @@
 #' @importFrom stats as.ts AIC BIC logLik nobs fitted residuals coef time
 NULL
+
+#' @importFrom zoo index
+NULL
+
+#' @importFrom generics tidy augment glance
+NULL