+++ R/reexports.R
@@ -1,16 +1,3 @@
-#' Objects exported from other packages
-#' @export
-#' @importFrom broom tidy
-broom::tidy
-
-#' @export
-#' @importFrom broom augment
-broom::augment
-
-#' @export
-#' @importFrom broom glance
-broom::glance
-
 #' @export
 #' @importFrom vctrs vec_ptype2
 vctrs::vec_ptype2
@@ -23,6 +10,4 @@ vctrs::vec_cast
 #' @importFrom xts as.xts
 as.ts.xts <- xts:::as.ts.xts
 
-#' @export
-#' @importFrom zoo index
-zoo::index
+